Milestone4/requirements.txt
google-genai
flask
playwright
reportlab